refactor(state): rename slice into feature
