+++ src/app/state/auth/auth.actions.ts
@@ -1,7 +1,7 @@
 import { createActionGroup, emptyProps, props } from '@ngrx/store';
 
 import { LoginCredentials } from 'app/core/models/auth/login.model';
-import { AuthSession } from './auth.slice';
+import { AuthSession } from './auth.feature';
 import { SignupCredentials } from 'app/core/models/auth/signup.model';
 import { User } from 'app/core/models/user/user.model';
