+++ src/app/state/auth/auth.effects.ts
@@ -7,7 +7,7 @@ import { catchError, map, of, switchMap, tap } from 'rxjs';
 import { AuthService } from 'app/core/services/auth/auth.service';
 import { AuthActions } from './auth.actions';
 import { User } from 'app/core/models/user/user.model';
-import { AuthSession } from './auth.slice';
+import { AuthSession } from './auth.feature';
 import { UserRole } from 'app/core/models/user/user-role.model';
 
 @Injectable()
