+++ src/app/state/auth/auth.feature.ts
@@ -14,7 +14,7 @@ export type AuthSessionState = AuthSession | undefined;
 
 const initialState: AuthSessionState = undefined;
 
-export const authSlice = createFeature({
+export const authFeature = createFeature({
   name: 'auth',
   reducer: createReducer(
     initialState as AuthSessionState,
